:truck: style: move commands to folder
