+++ thipstercli/commands/__init__.py
@@ -1,0 +1,3 @@
+"""THipster CLI commands"""
+from .providers import providers_app
+from .repository import repository_app
