+++ thipstercli/commands/providers.py
@@ -10,15 +10,15 @@ from thipstercli.helpers import (
     get_thipster_module_class_list,
 )
 
-provider_app = typer.Typer(no_args_is_help=True)
+providers_app = typer.Typer(no_args_is_help=True)
 
 
-@provider_app.callback()
+@providers_app.callback()
 def _callback():
     """Manage authentification providers."""
 
 
-@provider_app.command('list')
+@providers_app.command('list')
 def _list():
     """List all the supported providers."""
     state['providers'] = get_thipster_module_class_list('auth')
@@ -29,7 +29,7 @@ def _list():
     __more_info_provider()
 
 
-@provider_app.command('info')
+@providers_app.command('info')
 def info(provider: str):
     """Get information about a provider."""
     provider = check_provider_exists(provider)
@@ -38,7 +38,7 @@ def info(provider: str):
     print(Panel(provider_class.__doc__, title=provider))
 
 
-@provider_app.command('set')
+@providers_app.command('set')
 def set_auth_provider(provider: str):
     """Set the provider to use for the auth."""
     provider = check_provider_exists(provider)
@@ -51,7 +51,7 @@ def set_auth_provider(provider: str):
     __more_info_provider()
 
 
-@provider_app.command('display')
+@providers_app.command('display')
 def display():
     """Display the current provider."""
     if not state.get('auth_provider', None):
@@ -80,4 +80,4 @@ def __more_info_provider():
 
 
 if __name__ == '__main__':
-    provider_app()
+    providers_app()
